Исправления

+++ LibraryCPP/dllist.cpp
@@ -11,14 +11,12 @@ struct DLListItem
 struct DLList
 {
     DLListItem* head;
-    size_t length;
 };
 
 DLList* list_create()
 {
     DLList* list = new DLList;
     list->head = nullptr;
-    list->length = 0;
     return list;
 }
 
@@ -26,7 +24,7 @@ void list_delete(DLList* list)
 {
     DLListItem* current = list->head;
 
-    while (list->length != 0)
+    while (current != nullptr)
     {
         DLListItem* prev = current;
         current = list_item_next(prev);
@@ -61,7 +59,7 @@ DLListItem* list_insert(DLList* list, Data data)
     DLListItem* item = new DLListItem;
     item->data = data;
 
-    if (list->head == nullptr || list->length == 0)
+    if (list->head == nullptr)
     {
         list->head = item;
         list->head->next = item;
@@ -75,7 +73,6 @@ DLListItem* list_insert(DLList* list, Data data)
         list->head->prev = item;
     }
     list->head = item;
-    list->length++;
 
     return item;
 }
@@ -92,7 +89,6 @@ DLListItem* list_insert_after(DLList* list, DLListItem* item, Data data)
         item->next->prev = new_item;
     }
     item->next = new_item;
-    list->length++;
 
     return new_item;
 }
@@ -102,12 +98,14 @@ DLListItem* list_erase_first(DLList* list)
     DLListItem* delete_item = list->head;
     DLListItem* new_head = list->head->next;
 
-    delete_item->prev->next = new_head;
-    new_head->prev = delete_item->prev;
-    list->head = new_head;
+    if (new_head != nullptr)
+    {
+        delete_item->prev->next = new_head;
+        new_head->prev = delete_item->prev;
+        list->head = new_head;
+    }
 
     delete delete_item;
-    list->length--;
 
     return nullptr;
 }
@@ -116,21 +114,14 @@ DLListItem* list_erase(DLList* list, DLListItem* item)
 {
     DLListItem* item_prev = item->prev;
     DLListItem* item_next = item->next;
-    if (item_next != nullptr)
-    {
-        item_next->prev = item_prev;
-    }
-    if (item_prev != nullptr)
-    {
-        item_prev->next = item->next;
-    }
-
+    item_next->prev = item_prev;
+    item_prev->next = item->next;
+    
     if (item == list->head)
     {
         return list_erase_first(list);
     }
     delete item;
-    list->length--;
 
     return item_next;
 }
@@ -140,7 +131,3 @@ DLListItem* list_erase_next(DLList* list, DLListItem* item)
     return list_erase(list, item->next);
 }
 
-size_t list_get_length(const DLList* list)
-{
-    return list->length;
-}
